Disable keystore for beats in FIPS mode
